+++ bin/cfn/cluster-deploy
@@ -16,6 +16,7 @@ PARAMETERS=$(cfn-toml params v2 -t $CONFIG_PATH)
 aws cloudformation deploy \
     --stack-name $STACK_NAME \
     --s3-bucket $BUCKET \
+    --s3-prefix cluster \
     --region $REGION \
     --template-file "$CFN_PATH" \
     --no-execute-changeset \
