+++ bin/cfn/ddb-deploy
@@ -1,0 +1,17 @@
+#! /bin/env bash
+set -e #stop execution of the  script if it fails
+
+#PACKAGED_TEMPLATE_PATH="/workspace/aws-bootcamp-cruddur-2023/.aws-sam/build/packaged.yaml"
+PACKAGED_TEMPLATE_PATH="/Users/charlesfelteau/aws-bootcamp-cruddur-2023/.aws-sam/build/packaged.yaml"
+#CONFIG_PATH="/workspace/aws-bootcamp-cruddur-2023/aws/cfn/ddb/config.toml"
+CONFIG_PATH="/Users/charlesfelteau/aws-bootcamp-cruddur-2023/aws/cfn/ddb/config.toml"
+
+echo "=====> deploy"
+# https://docs.aws.amazon.com/serverless-application-model/latest/developerguide/sam-cli-command-reference-sam-deploy.html
+sam deploy \
+  --template-file $PACKAGED_TEMPLATE_PATH  \
+  --config-file $CONFIG_PATH \
+  --stack-name "CrdDdb" \
+  --tags group=cruddur-ddb \
+  --no-execute-changeset \
+  --capabilities "CAPABILITY_NAMED_IAM"
