+++ ddb/package-2
@@ -1,0 +1,19 @@
+#! /usr/bin/env bash
+set -e # stop the execution of the script if it fails
+
+ARTIFACT_BUCKET="cfn-artifacts-cfelt"
+#TEMPLATE_PATH="/workspace/aws-bootcamp-cruddur-2023/.aws-sam/build/template.yaml"
+TEMPLATE_PATH="/Users/charlesfelteau/aws-bootcamp-cruddur-2023/.aws-sam/build/template.yaml"
+#OUTPUT_TEMPLATE_PATH="/workspace/aws-bootcamp-cruddur-2023/.aws-sam/build/packaged.yaml"
+OUTPUT_TEMPLATE_PATH="/Users/charlesfelteau/aws-bootcamp-cruddur-2023/.aws-sam/build/packaged.yaml"
+#CONFIG_PATH="/workspace/aws-bootcamp-cruddur-2023/aws/cfn/ddb/config.toml"
+CONFIG_PATH="/Users/charlesfelteau/aws-bootcamp-cruddur-2023/ddb/config.toml"
+
+echo "=====> package"
+# https://docs.aws.amazon.com/serverless-application-model/latest/developerguide/sam-cli-command-reference-sam-package.html
+sam package \
+  --s3-bucket $ARTIFACT_BUCKET \
+  --config-file $CONFIG_PATH \
+  --output-template-file $OUTPUT_TEMPLATE_PATH \
+  --template-file $TEMPLATE_PATH \
+  --s3-prefix "ddb"
